+++ CHANGELOG.md
@@ -25,9 +25,10 @@ Changes are kept under subsections:
 Only major refactoring improvements or functional changes should be noted.
 
 --------------------------------------------------------------------------------
-## [1.4.0-bulk] - 2025-02-xx - Refactor Events Requests
+## [1.4.1-bulk] - 2025-02-xx - Refactor Events Requests
 #### Fixed
 #### Changed
+- NConf calculation optimized
 #### New
 
 ## [1.3.3] - 2025-02-07 - Partial bulk by CLI